rename: simplify names in `src/auth/credentials`
