+++ src/auth/providers/sharedCredentialsProvider.ts
@@ -13,7 +13,7 @@ import { fromEnv } from '@aws-sdk/credential-provider-env'
 import { getLogger } from '../../shared/logger'
 import { getStringHash } from '../../shared/utilities/textUtilities'
 import { getMfaTokenFromUser } from '../credentials/credentialsCreator'
-import { resolveProviderWithCancel } from '../credentials/credentialsUtilities'
+import { resolveProviderWithCancel } from '../credentials/utils'
 import { CredentialsProvider, CredentialsProviderType, CredentialsId } from './credentials'
 import { CredentialType } from '../../shared/telemetry/telemetry.gen'
 import { assertHasProps, getMissingProps, hasProps } from '../../shared/utilities/tsUtils'
