+++ src/auth/wizards/templates.ts
@@ -12,7 +12,7 @@ import { ProfileTemplateProvider } from './createProfile'
 import { createCommonButtons } from '../../shared/ui/buttons'
 import { credentialHelpUrl } from '../../shared/constants'
 import { SharedCredentialsKeys, StaticProfile } from '../credentials/types'
-import { getCredentialError } from '../credentials/sharedCredentialsValidation'
+import { getCredentialError } from '../credentials/validation'
 
 function getTitle(profileName: string): string {
     return localize('AWS.title.createCredentialProfile', 'Creating new profile "{0}"', profileName)
