+++ src/shared/sam/sync.ts
@@ -35,7 +35,7 @@ import { SamConfig } from './config'
 import { cast, Instance, Optional, Union } from '../utilities/typeConstructors'
 import { pushIf, toRecord } from '../utilities/collectionUtils'
 import { Auth, IamConnection } from '../../auth/auth'
-import { asEnvironmentVariables } from '../../auth/credentials/credentialsUtilities'
+import { asEnvironmentVariables } from '../../auth/credentials/utils'
 import { SamCliInfoInvocation } from './cli/samCliInfo'
 import { parse } from 'semver'
 import { isAutomation } from '../vscode/env'
