+++ src/test/fakeExtensionContext.ts
@@ -6,7 +6,7 @@
 import { mkdirp } from 'fs-extra'
 import * as vscode from 'vscode'
 import * as path from 'path'
-import { CredentialsStore } from '../auth/credentials/credentialsStore'
+import { CredentialsStore } from '../auth/credentials/store'
 import { ExtContext } from '../shared/extensions'
 import { SamCliContext } from '../shared/sam/cli/samCliContext'
 import {
